requirements.txt
requests>=2.31.0
websocket-client>=1.6.4
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.0
ccxt>=4.1.0
openai>=1.3.0
anthropic>=0.7.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
loguru>=0.7.0
alpaca-py>=0.20.0 